Moved compiled_simple_gettable_attrs() from genutils.importutils to genutils.attrs; updated dependent imports
