+++ polymerist/genutils/importutils.py
@@ -4,29 +4,13 @@ import logging
 LOGGER = logging.getLogger(__name__)
 
 from types import ModuleType
-from typing import Any, Generator, Iterable, Optional
+from typing import Generator, Iterable, Optional
 
-import re
 import pkgutil
 import importlib
 
 from enum import StrEnum
 from itertools import chain
-
-
-# ATTRIBUTE GETTING AND SETTING
-def compile_simple_getable_attrs(obj : Any, getter_str : str='get', repl_str : Optional[str]=None) -> dict[str, Any]:
-    '''Takes an object and returns a dict of the return values of all argument-free methods of the objects
-    Looks for methods of the object whose names contain with "getter_str", and can replace this with the value of "repl_str" in the final dict output if provided'''
-    getable_dict = {}
-    for attr_name in dir(obj):
-        if re.search(getter_str, attr_name): 
-            try:
-                attr_key = attr_name if (repl_str is None) else re.sub(getter_str, repl_str, attr_name)
-                getable_dict[attr_key] = getattr(obj, attr_name)()
-            except (TypeError, Exception): # TODO : find way to selectively intercept the Boost C++ wrapper ArgumentError
-                pass
-    return getable_dict
 
 
 # FILETREE CHARACTERS
